Changed the method to detect chat message language.  Separates modules.  Fixed the issue on member-only streamings.
