+++ src/manifest.chrome.json
@@ -1,7 +1,7 @@
 {
 	"manifest_version": 3,
 	"name": "__MSG_manifestName__",
-	"version": "0.14.1",
+	"version": "0.14.2",
 	"author": "_y_s",
 	"action": {
 		"default_state": "disabled"
